If checks to make sure links don't already exist; fix syntax error with variable reference

+++ setup.sh
@@ -3,7 +3,7 @@
 shopt -s extglob
 
 mkdir -p $HOME/.vim
-cp -vuR $(pwd)/vim/* $HOME/.vim
+cp -vuR $(pwd)/vim/* $HOME/.vim/
 cp -vu $(pwd)/vimrc $HOME/.vimrc
 curl -fLo $HOME/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
@@ -12,12 +12,17 @@ cp -vu $(pwd)/tmux.conf $HOME/.tmux.conf
 
 if hash nvim 2>/dev/null; then
     echo "Found neovim installation, adding neovim specific configs"
-    mkdir -p $HOME/.config
-    cp -vuR $(pwd)/nvim/* $HOME/.config/nvim
-    ln -vs $HOME/.vimrc $HOME/.config/nvim/init.vim
+    mkdir -p $HOME/.config/nvim
+    cp -vuR $(pwd)/nvim/* $HOME/.config/nvim/
+    if [ ! -s $HOME/.config/nvim/init.vim ]; then
+        ln -vs $HOME/.vimrc $HOME/.config/nvim/init.vim
+    fi
+
     mkdir -p $HOME/.local/share/nvim/site/autoload
-    cp $HOME/.vim/autoload/plug.vim $HOME/.local/share/nvim/site/autoload/plug.vim
-    ln -vs $HOME/.vim/plugin $HOME/.local/share/nvim/site/plugin
+    cp -u $HOME/.vim/autoload/plug.vim $HOME/.local/share/nvim/site/autoload/plug.vim
+    if [ ! -s $HOME/.local/share/nvim/site/plugin ]; then
+        ln -vs $HOME/.vim/plugin $HOME/.local/share/nvim/site/plugin
+    fi
 else
     echo "Neovim not found, skipping neovim configs"
 fi
@@ -30,7 +35,9 @@ if hash git 2>/dev/null; then
     fi
     echo "Symlinking prezto runcoms..."
     for rcfile in "${ZDOTDIR:-$HOME}"/.zprezto/runcoms/!(README.md); do
-        ln -vs "$rcfile" "${ZDOTDIR:-$HOME}/.${rcfile##*/}"
+        if [ ! -s "${ZDOTDIR:-$HOME}/.${rcfile##*/}" ]; then
+            ln -vs "$rcfile" "${ZDOTDIR:-$HOME}/.${rcfile##*/}"
+        fi
     done
     if hash zsh 2>/dev/null; then
         sudo chsh -s $(which zsh) ${SUDO_USER:-$USER}
@@ -42,13 +49,13 @@ else
 fi
 
 install_solarized_dircolors = 0
-read -p "Install solarized dircolors? [y/N]" ans
+read -p "Install solarized dircolors? [y/N] " ans
 case "$ans" in
     [Yy]|[Yy][Ee][Ss] ) install_solarized_dircolors = 1 ;;
     * ) break ;;
 esac
 
-if [ install_solarized_dircolors -eq 1 ]; then
+if [ $install_solarized_dircolors -eq 1 ]; then
     git clone --quiet https://github.com/seebi/dircolors-solarized.git
     cp -uv dircolors-solarized/dircolors.256dark $HOME/.dir_colors
 fi
